+++ commons/src/main/java/org/dhis2/commons/resources/MetadataIconProvider.kt
@@ -17,7 +17,11 @@ import java.io.File
 
 class MetadataIconProvider(private val d2: D2) {
 
-    operator fun invoke(color: String?, icon: String?, defaultColor: Color = SurfaceColor.Primary): MetadataIconData {
+    operator fun invoke(
+        color: String?,
+        icon: String?,
+        defaultColor: Color = SurfaceColor.Primary,
+    ): MetadataIconData {
         val objectStyleBuilder = ObjectStyle.builder()
         color?.let { objectStyleBuilder.color(it) }
         icon?.let { objectStyleBuilder.icon(it) }
@@ -26,37 +30,38 @@ class MetadataIconProvider(private val d2: D2) {
 
     operator fun invoke(style: ObjectStyle) = invoke(style, SurfaceColor.Primary)
 
-    operator fun invoke(style: ObjectStyle, defaultColor: Color?) = style.icon()?.let {
-        d2.iconModule().icons().key(it).blockingGet()
-    }.let { icon ->
-        val imageCardData = when {
-            (icon is Icon.Custom) -> {
-                provideImageBitmap(icon.path)?.let { imageBitmap ->
-                    ImageCardData.CustomIconData(
+    operator fun invoke(style: ObjectStyle, defaultColor: Color? = SurfaceColor.Primary) =
+        style.icon()?.let {
+            d2.iconModule().icons().key(it).blockingGet()
+        }.let { icon ->
+            val imageCardData = when {
+                (icon is Icon.Custom) -> {
+                    provideImageBitmap(icon.path)?.let { imageBitmap ->
+                        ImageCardData.CustomIconData(
+                            uid = "",
+                            label = "",
+                            image = imageBitmap,
+                        )
+                    } ?: ImageCardData.IconCardData(
                         uid = "",
                         label = "",
-                        image = imageBitmap,
+                        iconRes = FILE_NOT_LOADED,
+                        iconTint = style.color()?.toColor() ?: defaultColor ?: Color.Unspecified,
                     )
-                } ?: ImageCardData.IconCardData(
+                }
+
+                else -> ImageCardData.IconCardData(
                     uid = "",
                     label = "",
-                    iconRes = FILE_NOT_LOADED,
-                    iconTint = style.color()?.toColor() ?: defaultColor ?: Color.Unspecified,
+                    iconRes = provideIconRes(style.icon()),
+                    iconTint = style.color()?.toColor() ?: Color.Unspecified,
                 )
             }
-
-            else -> ImageCardData.IconCardData(
-                uid = "",
-                label = "",
-                iconRes = provideIconRes(style.icon()),
-                iconTint = style.color()?.toColor() ?: Color.Unspecified,
+            MetadataIconData(
+                imageCardData = imageCardData,
+                color = style.color()?.toColor() ?: defaultColor ?: Color.Unspecified,
             )
         }
-        MetadataIconData(
-            imageCardData = imageCardData,
-            color = style.color()?.toColor() ?: defaultColor ?: Color.Unspecified,
-        )
-    }
 
     private fun provideIconRes(icon: String?): String =
         when {
